refactoring contactfragment into contactbasefragment class
